Sema: Move some code to a new file TypeCheckStorage.cpp
