+++ lib/Sema/CMakeLists.txt
@@ -58,6 +58,7 @@ add_swift_host_library(swiftSema STATIC
   TypeCheckREPL.cpp
   TypeCheckRequestFunctions.cpp
   TypeCheckStmt.cpp
+  TypeCheckStorage.cpp
   TypeCheckSwitchStmt.cpp
   TypeCheckType.cpp
   TypeChecker.cpp
